feat(api): Add startApi entrypoint for CLI imports
Extract bootstrap logic into start.ts, returning the Nest app instance so the CLI can manage lifecycle. Refactor database module to decouple MongoMemoryServer into mongo-uri.ts, supporting an externally injected URI. Update package.json with metadata, exports, and fields for npm publishing (MIT license, prebuilt dist).
