+++ packages/mukti-api/package.json
@@ -1,10 +1,36 @@
 {
   "name": "@mukti/api",
   "version": "0.1.1-beta.0",
-  "description": "",
-  "author": "",
-  "private": true,
-  "license": "UNLICENSED",
+  "description": "Mukti API — the NestJS backend behind `npx muktiai`, published prebuilt.",
+  "author": "Mukti",
+  "license": "MIT",
+  "homepage": "https://mukti.chat",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/shettydev/mukti.git",
+    "directory": "packages/mukti-api"
+  },
+  "bugs": {
+    "url": "https://github.com/shettydev/mukti/issues"
+  },
+  "engines": {
+    "node": ">=20.11"
+  },
+  "main": "./dist/start.js",
+  "types": "./dist/start.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/start.d.ts",
+      "default": "./dist/start.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
   "scripts": {
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
@@ -79,7 +105,6 @@
     "eslint-plugin-prettier": "^5.2.2",
     "globals": "^16.0.0",
     "jest": "^29.7.0",
-    "mongodb-memory-server": "^11.2.0",
     "prettier": "^3.4.2",
     "source-map-support": "^0.5.21",
     "supertest": "^7.0.0",
@@ -89,6 +114,9 @@
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.20.0"
+  },
+  "optionalDependencies": {
+    "mongodb-memory-server": "^11.2.0"
   },
   "jest": {
     "moduleFileExtensions": [
